feat(widget): refactored main hook and separated code to bussinnes logic parts
