+++ src/components/screens/widget/useAnyaltWidget.ts
@@ -1,49 +1,31 @@
-import { useBitcoinWallet } from '@ant-design/web3-bitcoin';
-import { BestRouteResponse, SupportedToken } from '@anyalt/sdk';
+import { BestRouteResponse } from '@anyalt/sdk';
 import { useDisclosure, useSteps } from '@chakra-ui/react';
-import { useWallet } from '@solana/wallet-adapter-react';
-import { useAtom, useAtomValue } from 'jotai';
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useAccount } from 'wagmi';
+import { useAtom, useAtomValue, useSetAtom } from 'jotai';
+import { useCallback, useEffect, useMemo } from 'react';
 import {
-  ChainType,
   EstimateResponse,
   Token,
   WalletConnector,
   WidgetTemplateType,
 } from '../../..';
 import {
-  DEBOUNCE_TIMEOUT,
-  REFRESH_INTERVAL,
-} from '../../../constants/transaction';
-import {
   activeOperationIdAtom,
-  allChainsAtom,
-  anyaltInstanceAtom,
   bestRouteAtom,
-  currentStepAtom,
-  lastMileTokenEstimateAtom,
   selectedRouteAtom,
   selectedTokenAmountAtom,
   selectedTokenAtom,
   showStuckTransactionDialogAtom,
-  slippageAtom,
-  swapDataAtom,
-  swapResultTokenAtom,
-  tokenFetchErrorAtom,
   transactionIndexAtom,
-  transactionsListAtom,
   transactionsProgressAtom,
 } from '../../../store/stateStore';
 import { TransactionsProgress } from '../../../types/transaction';
-import {
-  calculateWorstOutput,
-  convertSwapTransactionToTransactionProgress,
-} from '../../../utils';
-import { ChainIdToChainConstant } from '../../../utils/chains';
+import { convertSwapTransactionToTransactionProgress } from '../../../utils';
 import { usePendingOperation } from '../../standalones/pendingOperationDialog/usePendingOperation';
-import { useTokenInputBox } from '../../standalones/selectSwap/token/input/useTokenInputBox';
+import { useConfirmRoute } from './useConfirmRoute';
+import { useControllWidget } from './useControllWidget';
+import { useFetchRoutes } from './useFetchRoutes';
 import { useSetupWidget } from './useSetupWidget';
+import { useWidgetWallets } from './useWidgetWallets';
 
 type Props = {
   estimateCallback: (token: Token) => Promise<EstimateResponse>;
@@ -56,6 +38,33 @@ type Props = {
   onClose: () => void;
 };
 
+type ReturnType = {
+  loading: boolean;
+  activeStep: number;
+  activeRoute: BestRouteResponse | undefined;
+  isValidAmountIn: boolean;
+  isButtonDisabled: boolean;
+  openSlippageModal: boolean;
+  isConnectWalletsOpen: boolean;
+  failedToFetchRoute: boolean;
+  areWalletsConnected: boolean;
+  showPendingOperationDialog: boolean;
+  showStuckTransactionDialog: boolean;
+  allNecessaryWalletsConnected: boolean;
+  modalWrapperMaxWidth: string | undefined;
+  headerCustomText: string | undefined;
+  onBackClick: () => void;
+  onComplete: () => void;
+  onTxComplete: () => void;
+  onConfigClick: () => void;
+  connectWalletsOpen: () => void;
+  connectWalletsClose: () => void;
+  setOpenSlippageModal: (value: boolean) => void;
+  onChooseRouteButtonClick: () => Promise<void>;
+  setOperationToCurrentRoute: (operation: BestRouteResponse) => void;
+  resetState: () => void;
+};
+
 export const useAnyaltWidget = ({
   apiKey,
   swapResultToken,
@@ -65,27 +74,28 @@ export const useAnyaltWidget = ({
   widgetTemplate,
   estimateCallback,
   onClose,
-}: Props) => {
-  const [loading, setLoading] = useState(false);
-  const [isValidAmountIn, setIsValidAmountIn] = useState(true);
-  const [openSlippageModal, setOpenSlippageModal] = useState(false);
-  const [failedToFetchRoute, setFailedToFetchRoute] = useState(false);
+}: Props): ReturnType => {
+  const selectedToken = useAtomValue(selectedTokenAtom);
+  const selectedRoute = useAtomValue(selectedRouteAtom);
+  const selectedTokenAmount = useAtomValue(selectedTokenAmountAtom);
+  const showStuckTransactionDialog = useAtomValue(
+    showStuckTransactionDialogAtom,
+  );
 
-  const { address: evmAddress, isConnected: isEvmConnected } = useAccount();
-  const { publicKey: solanaAddress, connected: isSolanaConnected } =
-    useWallet();
-  const { account: bitcoinAccount } = useBitcoinWallet();
+  const [bestRoute, setBestRoute] = useAtom(bestRouteAtom);
+
+  const setTransactionIndex = useSetAtom(transactionIndexAtom);
+  const setActiveOperationId = useSetAtom(activeOperationIdAtom);
+  const setTransactionsProgress = useSetAtom(transactionsProgressAtom);
+
   const { activeStep, setActiveStep, goToPrevious } = useSteps({
     index: 0,
   });
-
   const {
     isOpen: isConnectWalletsOpen,
     onClose: connectWalletsClose,
     onOpen: connectWalletsOpen,
   } = useDisclosure();
-
-  const { balance } = useTokenInputBox();
 
   const { showPendingOperationDialog, allNecessaryWalletsConnected } =
     usePendingOperation({ closeConnectWalletsModal: connectWalletsClose });
@@ -98,37 +108,63 @@ export const useAnyaltWidget = ({
     widgetTemplate,
   });
 
-  const slippage = useAtomValue(slippageAtom);
-  const allChains = useAtomValue(allChainsAtom);
-  const selectedRoute = useAtomValue(selectedRouteAtom);
-  const anyaltInstance = useAtomValue(anyaltInstanceAtom);
-  const swapResultTokenGlobal = useAtomValue(swapResultTokenAtom);
-  const showStuckTransactionDialog = useAtomValue(
-    showStuckTransactionDialogAtom,
-  );
+  const {
+    evmAddress,
+    solanaAddress,
+    bitcoinAccount,
+    areWalletsConnected,
+    getChain,
+  } = useWidgetWallets({
+    walletConnector,
+    isConnectWalletsOpen,
+    connectWalletsClose,
+  });
 
-  const [selectedToken, setSelectedToken] = useAtom(selectedTokenAtom);
-  const [selectedTokenAmount, setSelectedTokenAmount] = useAtom(
-    selectedTokenAmountAtom,
-  );
-  const [lastMileTokenEstimate, setLastMileTokenEstimate] = useAtom(
-    lastMileTokenEstimateAtom,
-  );
+  const {
+    loading,
+    isValidAmountIn,
+    failedToFetchRoute,
+    onGetRoutes,
+    setLoading,
+    setListOfTransactionsFromRoute,
+  } = useFetchRoutes({
+    finalToken,
+    activeStep,
+    swapResultToken,
+    minDepositAmount,
+    setActiveStep,
+    estimateCallback,
+  });
 
-  const [, setCurrentStep] = useAtom(currentStepAtom);
+  const {
+    openSlippageModal,
+    setOpenSlippageModal,
+    resetState,
+    onConfigClick,
+    onBackClick,
+    onTxComplete,
+    onComplete,
+  } = useControllWidget({
+    activeStep,
+    setActiveStep,
+    onGetRoutes,
+    goToPrevious,
+    onClose,
+  });
 
-  const [swapData, setSwapData] = useAtom(swapDataAtom);
-  const [bestRoute, setBestRoute] = useAtom(bestRouteAtom);
-  const [, setTokenFetchError] = useAtom(tokenFetchErrorAtom);
-  const [, setTransactionsList] = useAtom(transactionsListAtom);
-  const [, setTransactionIndex] = useAtom(transactionIndexAtom);
-  const [, setActiveOperationId] = useAtom(activeOperationIdAtom);
-
-  const [, setTransactionsProgress] = useAtom(transactionsProgressAtom);
-
-  useEffect(() => {
-    onGetQuote(false);
-  }, [selectedToken, slippage, balance]);
+  const { onChooseRouteButtonClick } = useConfirmRoute({
+    areWalletsConnected,
+    evmAddress,
+    solanaAddress,
+    bitcoinAccount,
+    widgetTemplate,
+    walletConnector,
+    connectWalletsOpen,
+    connectWalletsClose,
+    setActiveStep,
+    setLoading,
+    getChain,
+  });
 
   const isButtonDisabled = useMemo(() => {
     if (activeStep === 0) {
@@ -141,466 +177,10 @@ export const useAnyaltWidget = ({
     return Number(selectedTokenAmount ?? 0) == 0 || selectedToken == null;
   }, [selectedTokenAmount, selectedToken, bestRoute, activeStep]);
 
-  const resetState = useCallback(() => {
-    setActiveStep(0);
-    setActiveOperationId(undefined);
-    setLastMileTokenEstimate(undefined);
-    setTransactionsList(undefined);
-    setSelectedTokenAmount(undefined);
-    setSelectedToken(undefined);
-    setTokenFetchError({ isError: false, errorMessage: '' });
-    setBestRoute(undefined);
-    setSwapData({
-      swapIsFinished: false,
-      isCrosschainSwapError: false,
-      crosschainSwapOutputAmount: '',
-      totalSteps: 0,
-      currentStep: 1,
-    });
-    setTransactionsProgress({});
-    setTransactionIndex(1);
-    localStorage.removeItem('operationId');
-    localStorage.removeItem('tokenBuyOperationId');
-  }, [
-    setActiveOperationId,
-    setActiveOperationId,
-    setLastMileTokenEstimate,
-    setTransactionsList,
-    setSelectedTokenAmount,
-    setSelectedToken,
-    setTokenFetchError,
-    setBestRoute,
-    setSwapData,
-    setTransactionsProgress,
-    setTransactionIndex,
-  ]);
-
-  useEffect(() => {
-    setCurrentStep(activeStep);
-  }, [activeStep]);
-
-  //TODO: Should be triggered, once all routes has been setted. Also figure out how to handle for multiple routes.
-  useEffect(() => {
-    if (bestRoute) {
-      const token = {
-        ...swapResultToken,
-        amount: bestRoute.outputAmount.toString(),
-      };
-      estimateCallback(token).then((res) => {
-        setLastMileTokenEstimate(res);
-      });
-    }
-  }, [bestRoute]);
-
   //TODO: Should be refactored to handle it to handle selected route. Probably can be deleted
   useEffect(() => {
     if (selectedRoute) setBestRoute(selectedRoute);
   }, [selectedRoute]);
-
-  const setListOfTransactionsFromRoute = useCallback(
-    (route: BestRouteResponse, inputToken: Partial<SupportedToken>) => {
-      const lastStepOfOperation = route?.swapSteps[route?.swapSteps.length - 1];
-      const lastTokenOfOperation = lastStepOfOperation?.destinationToken;
-
-      setTransactionsList({
-        steps: [
-          ...(route?.swapSteps.map((step) => {
-            const toPayout = !isNaN(parseInt(step.payout))
-              ? step.payout
-              : step.quotePayout;
-            return {
-              from: {
-                tokenName: step.sourceToken.symbol,
-                tokenLogo: step.sourceToken.logo,
-                tokenAmount: step.amount,
-                tokenPrice: (
-                  parseFloat(step.amount) * step.sourceToken.tokenUsdPrice
-                ).toFixed(2),
-                tokenUsdPrice: step.sourceToken.tokenUsdPrice.toFixed(2),
-                tokenDecimals: step.sourceToken.decimals,
-                blockchain: step.sourceToken.blockchain,
-                blockchainLogo: step.sourceToken.blockchainLogo,
-              },
-              to: {
-                tokenName: step.destinationToken.symbol,
-                tokenLogo: step.destinationToken.logo,
-                tokenAmount: toPayout,
-                tokenPrice: (
-                  parseFloat(toPayout) * step.destinationToken.tokenUsdPrice
-                ).toFixed(2),
-                tokenUsdPrice: step.destinationToken.tokenUsdPrice.toFixed(2),
-                tokenDecimals: step.destinationToken.decimals,
-                blockchain: step.destinationToken.blockchain,
-                blockchainLogo: step.destinationToken.blockchainLogo,
-              },
-            };
-          }) || []),
-          {
-            from: {
-              tokenName: lastTokenOfOperation?.symbol || '',
-              tokenLogo: lastTokenOfOperation?.logo || '',
-              tokenAmount: lastStepOfOperation?.amount || '',
-              tokenPrice: lastStepOfOperation?.amount || '',
-              tokenUsdPrice:
-                lastTokenOfOperation?.tokenUsdPrice.toFixed(2) || '',
-              tokenDecimals: lastTokenOfOperation?.decimals || 0,
-              blockchain: lastTokenOfOperation?.blockchain || '',
-              blockchainLogo: lastTokenOfOperation?.blockchainLogo || '',
-            },
-            to: {
-              tokenName: finalToken?.name || '',
-              tokenLogo: finalToken?.logoUrl || '',
-              tokenAmount: lastMileTokenEstimate?.amountOut || '',
-              tokenPrice:
-                (
-                  parseFloat(lastMileTokenEstimate?.priceInUSD || '0') /
-                  parseFloat(lastMileTokenEstimate?.amountOut || '1')
-                ).toFixed(2) || '',
-              tokenUsdPrice: lastMileTokenEstimate?.priceInUSD || '0',
-              tokenDecimals: finalToken?.decimals || 0,
-              blockchain: inputToken.chain?.displayName || '',
-              blockchainLogo: inputToken.chain?.logoUrl || '',
-            },
-          },
-        ],
-      });
-    },
-    [lastMileTokenEstimate, finalToken, swapResultTokenGlobal],
-  );
-
-  const onGetQuote = async (withGoNext: boolean = true) => {
-    if (activeStep > 1) return;
-    if (!selectedToken || !swapResultTokenGlobal || !selectedTokenAmount)
-      return;
-
-    if (selectedToken.id === swapResultTokenGlobal.id) {
-      setBestRoute({
-        outputAmount: selectedTokenAmount,
-        swapSteps: [],
-        operationId: '',
-        missingWalletForSourceBlockchain: false,
-      });
-
-      setTokenFetchError({
-        isError: false,
-        errorMessage: '',
-      });
-      return;
-    }
-
-    try {
-      setLoading(true);
-
-      const route = await anyaltInstance?.getBestRoute({
-        fromToken: {
-          address: selectedToken.tokenAddress ?? '',
-          chainName: selectedToken.chainName,
-        },
-        toToken: {
-          address: swapResultToken.address,
-          chainName:
-            ChainIdToChainConstant[
-              swapResultToken.chainId! as keyof typeof ChainIdToChainConstant
-            ],
-        },
-        amount: selectedTokenAmount,
-        slippage,
-      });
-
-      //TODO: Instead of setting best route. It should set all routes.
-      setBestRoute(route);
-
-      if (route && swapResultToken) {
-        setListOfTransactionsFromRoute(route, swapResultToken);
-      }
-
-      const tokensOut = parseFloat(route?.outputAmount || '0');
-      let isEnoughDepositTokens = tokensOut > minDepositAmount;
-
-      setTokenFetchError({
-        isError: !isEnoughDepositTokens,
-        errorMessage: `Amount should be equal or greater than ${minDepositAmount} ${swapResultToken?.symbol}`,
-      });
-
-      if (isEnoughDepositTokens && route) {
-        const { humanReadable: worstCaseOutput } = calculateWorstOutput(
-          route,
-          slippage,
-        );
-
-        if (parseFloat(worstCaseOutput) < minDepositAmount) {
-          isEnoughDepositTokens = false;
-          setTokenFetchError({
-            isError: true,
-            errorMessage: `Output possibly low, the transaction might not get executed due to slippage. The protocol expects a minimum of ${minDepositAmount} ${swapResultToken?.symbol} please increase the input amount.`,
-          });
-        }
-      }
-
-      if (balance && parseFloat(balance) < parseFloat(selectedTokenAmount)) {
-        setTokenFetchError({
-          isError: true,
-          errorMessage: `You don't have enough tokens in your wallet.`,
-        });
-      }
-
-      setIsValidAmountIn(isEnoughDepositTokens);
-      setFailedToFetchRoute(false);
-      if (activeStep === 0) setActiveStep(1);
-      if (withGoNext && isEnoughDepositTokens) setActiveStep(1);
-    } catch (error) {
-      console.error(error);
-      setTokenFetchError({
-        isError: true,
-        errorMessage: 'No Available Route',
-      });
-      setFailedToFetchRoute(true);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  //TODO: Can be part of getting routes hook.
-  useEffect(() => {
-    //Show loading state immediatly instead of waiting for a delay
-    if (selectedTokenAmount && selectedToken) setLoading(true);
-
-    const debounceTimeout = setTimeout(() => {
-      if (selectedTokenAmount && selectedToken) {
-        onGetQuote(false);
-      }
-    }, DEBOUNCE_TIMEOUT);
-
-    return () => {
-      clearTimeout(debounceTimeout);
-    };
-  }, [selectedToken, swapResultTokenGlobal, selectedTokenAmount]);
-
-  //TODO: This code is realted to modal, can be move to another places
-  const onConfigClick = () => {
-    setOpenSlippageModal(true);
-  };
-
-  const onChooseRouteButtonClick = async () => {
-    try {
-      if (areWalletsConnected) {
-        await confirmSelectedRoute();
-        setTransactionsProgress({});
-      } else {
-        if (walletConnector) {
-          walletConnector.connect();
-        } else {
-          connectWalletsOpen();
-        }
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  const getChain = (blockchain: string) =>
-    allChains.find((chain) => chain.name === blockchain);
-
-  const confirmSelectedRoute = async () => {
-    try {
-      setLoading(true);
-
-      let destination = '';
-
-      const selectedWallets: Record<string, string> = {};
-      //TODO: It shoudl read data from selected route.
-      bestRoute?.swapSteps.forEach((swapStep, index) => {
-        const fromBlockchain = swapStep.sourceToken.blockchain;
-        const toBlockchain = swapStep.destinationToken.blockchain;
-
-        const isSolanaFrom = fromBlockchain === 'SOLANA';
-        const isSolanaTo = toBlockchain === 'SOLANA';
-        const isBitcoinFrom = fromBlockchain === 'BTC';
-        const isBitcoinTo = toBlockchain === 'BTC';
-
-        if (isSolanaFrom || isSolanaTo) {
-          selectedWallets['SOLANA'] = solanaAddress?.toString() || '';
-        }
-        if (isBitcoinFrom || isBitcoinTo) {
-          selectedWallets['BTC'] = bitcoinAccount?.address || '';
-        }
-
-        const fromChain = getChain(fromBlockchain);
-        const toChain = getChain(toBlockchain);
-
-        const isEvmFrom = fromChain?.chainType === ChainType.EVM;
-        const isEvmTo = toChain?.chainType === ChainType.EVM;
-
-        if (isEvmFrom || isEvmTo) {
-          if (!evmAddress) {
-            throw new Error('EVM Wallet not connected');
-          }
-        }
-
-        if (isSolanaFrom || isSolanaTo) {
-          if (!solanaAddress) {
-            throw new Error('Solana Wallet not connected');
-          }
-        }
-
-        if (isBitcoinFrom || isBitcoinTo) {
-          if (!bitcoinAccount) {
-            throw new Error('Bitcoin Wallet not connected');
-          }
-        }
-
-        if (isEvmFrom) selectedWallets[fromBlockchain] = evmAddress!;
-        if (isEvmTo) selectedWallets[toBlockchain] = evmAddress!;
-
-        if (index === bestRoute.swapSteps.length - 1) {
-          switch (true) {
-            case isEvmTo:
-              destination = evmAddress!;
-              break;
-            case isSolanaTo:
-              destination = solanaAddress!.toString();
-              break;
-            case isBitcoinTo:
-              destination = bitcoinAccount!.address;
-              break;
-            default:
-              throw new Error('Destination not found');
-          }
-        }
-      });
-
-      if (bestRoute?.swapSteps.length === 0) {
-        const res = await anyaltInstance?.createOperation();
-        if (!res?.operationId) throw new Error('Failed to create operation');
-
-        setActiveOperationId(res?.operationId);
-      } else {
-        if (!bestRoute?.operationId) return;
-        const res = await anyaltInstance?.confirmRoute({
-          operationId: bestRoute?.operationId,
-          selectedWallets,
-          destination: destination,
-        });
-
-        if (!res?.operationId || !res?.ok)
-          throw new Error('Failed to confirm route');
-
-        setActiveOperationId(res?.operationId);
-        setBestRoute(bestRoute);
-        const localStorageKey =
-          widgetTemplate === 'TOKEN_BUY'
-            ? 'tokenBuyOperationId'
-            : 'operationId';
-        localStorage.setItem(localStorageKey, res.operationId);
-      }
-
-      connectWalletsClose();
-      setActiveStep(2);
-    } catch (error) {
-      console.error(error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const onBackClick = () => {
-    if (activeStep === 2) {
-      setActiveStep(1);
-      onGetQuote(false);
-      setSwapData({
-        ...swapData,
-        swapIsFinished: false,
-        isCrosschainSwapError: false,
-      });
-    } else {
-      goToPrevious();
-    }
-  };
-
-  const onTxComplete = () => {
-    setActiveStep(3);
-  };
-
-  useEffect(() => {
-    if (activeStep === 1 && bestRoute) {
-      const interval = setInterval(() => {
-        // Capture latest values inside the interval callback
-        const currentInToken = selectedToken;
-        const currentProtocolInputToken = swapResultTokenGlobal;
-        const currentInTokenAmount = selectedTokenAmount;
-        const userSelectedToken =
-          currentInToken &&
-          currentProtocolInputToken &&
-          currentInTokenAmount &&
-          bestRoute;
-
-        if (userSelectedToken) {
-          onGetQuote(false);
-        }
-      }, REFRESH_INTERVAL);
-
-      return () => clearInterval(interval);
-    }
-  }, [bestRoute]);
-
-  const areWalletsConnected = useMemo(() => {
-    let isSolanaRequired = false;
-    let isEvmRequired = false;
-    let isBitcoinRequired = false;
-
-    if (walletConnector && walletConnector.isConnected) {
-      return walletConnector.isConnected;
-    }
-
-    // Set chain flags for last mile tx
-    if (swapResultTokenGlobal?.chain?.chainType === ChainType.EVM) {
-      isEvmRequired = true;
-    } else if (swapResultTokenGlobal?.chain?.chainType === ChainType.SOLANA) {
-      isSolanaRequired = true;
-    }
-
-    bestRoute?.swapSteps.forEach((swapStep) => {
-      const fromBlockchain = swapStep.sourceToken.blockchain;
-      const toBlockchain = swapStep.destinationToken.blockchain;
-      const isSolanaFrom = fromBlockchain === 'SOLANA';
-      const isSolanaTo = toBlockchain === 'SOLANA';
-      const isBitcoinFrom = fromBlockchain === 'BTC';
-      const isBitcoinTo = toBlockchain === 'BTC';
-
-      if (isSolanaFrom || isSolanaTo) isSolanaRequired = true;
-      if (isBitcoinFrom || isBitcoinTo) isBitcoinRequired = true;
-
-      const fromChain = getChain(fromBlockchain);
-      const toChain = getChain(toBlockchain);
-
-      if (
-        fromChain?.chainType === ChainType.EVM ||
-        toChain?.chainType === ChainType.EVM
-      ) {
-        isEvmRequired = true;
-      }
-    });
-
-    let isWalletConnected = true;
-    if (isSolanaRequired && !isSolanaConnected) isWalletConnected = false;
-    if (isEvmRequired && !isEvmConnected) isWalletConnected = false;
-    if (isBitcoinRequired && !bitcoinAccount) isWalletConnected = false;
-
-    return isWalletConnected;
-  }, [isSolanaConnected, isEvmConnected, bestRoute, bitcoinAccount]);
-
-  useEffect(() => {
-    if (areWalletsConnected && isConnectWalletsOpen) {
-      connectWalletsClose();
-    }
-  }, [areWalletsConnected]);
-
-  const onComplete = () => {
-    onClose();
-    setActiveStep(0);
-    setTransactionIndex(1);
-    resetState();
-  };
 
   const setOperationToCurrentRoute = useCallback(
     (operation: BestRouteResponse) => {
